+++ cpu/idt.h
@@ -34,7 +34,7 @@ idt_register_t idt_reg;
 
 
 /* Functions implemented in idt.c */
-void set_idt_gate(int n, uint32_t handler);
+void set_idt_gate(int n, uint64_t handler);
 void set_idt();
 
 #endif
